requirements.txt
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pyOpenSSL==25.0.0
pytest==7.4.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
urllib3==1.26.20
urllib3-secure-extra==0.1.0
wsproto==1.2.0
